chore: done setup route and state
