+++ store/global/user.ts
@@ -1,17 +1,5 @@
 import { StateCreator } from 'zustand'
-import { GlobalStore } from '.'
-
-// TODO: To be defined
-interface UserInfoModel {
-  id: string
-}
-
-export interface UserSlice {
-  isLoggedIn: boolean
-  userInfo: UserInfoModel | null
-  setUserInfo: (userInfo: UserInfoModel) => void
-  setIsLoggedIn: (isLoggedIn: boolean) => void
-}
+import { GlobalStore, UserSlice } from './shared.types'
 
 export const createUserSlice: StateCreator<GlobalStore, [['zustand/persist', unknown]], [], UserSlice> = (set) => ({
   isLoggedIn: false,
